fix AI review; move format helper to a header file
