+++ components/thread_pool/include/thread_pool.hpp
@@ -119,16 +119,4 @@ private:
 
 } // namespace espp
 
-// fmt formatter for espp::ThreadPool::Stats
-template <> struct fmt::formatter<espp::ThreadPool::Stats> {
-  template <typename ParseContext> constexpr auto parse(ParseContext &ctx) const {
-    return ctx.begin();
-  }
-
-  template <typename FormatContext>
-  auto format(espp::ThreadPool::Stats const &s, FormatContext &ctx) const {
-    return fmt::format_to(ctx.out(),
-                          "ThreadPool::Stats{{submitted: {}, executed: {}, rejected: {}}}",
-                          s.submitted, s.executed, s.rejected);
-  }
-};
+#include "thread_pool_format_helpers.hpp"
